Book now decrements
When clicking the add button the count for the said book is dropped by
one. Implementing the thank you page next to allow the user to see the
request one more time and then put the user on his profile where it
shows the book he rented and the option to return.